refactor: 날씨 생성, 갱신 등 관련 메서드 리팩토링
