+++ back/src/main/java/org/pknu/weather/service/WeatherService.java
@@ -1,8 +1,9 @@
 package org.pknu.weather.service;
 
+import static org.pknu.weather.dto.converter.ExtraWeatherConverter.toExtraWeather;
+import static org.pknu.weather.dto.converter.ExtraWeatherConverter.toExtraWeatherInfo;
 import static org.pknu.weather.dto.converter.LocationConverter.toLocationDTO;
 
-import jakarta.persistence.EntityManager;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -10,75 +11,35 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.pknu.weather.apiPayload.code.status.ErrorStatus;
-import org.pknu.weather.common.WeatherParamsFactory;
-import org.pknu.weather.common.formatter.DateTimeFormatter;
-import org.pknu.weather.common.utils.GeometryUtils;
 import org.pknu.weather.domain.ExtraWeather;
 import org.pknu.weather.domain.Location;
 import org.pknu.weather.domain.Member;
 import org.pknu.weather.domain.Weather;
-import org.pknu.weather.dto.WeatherApiResponse;
 import org.pknu.weather.dto.WeatherResponse;
 import org.pknu.weather.dto.WeatherResponse.ExtraWeatherInfo;
 import org.pknu.weather.exception.GeneralException;
-import org.pknu.weather.feignClient.WeatherFeignClient;
-import org.pknu.weather.feignClient.dto.PointDTO;
-import org.pknu.weather.feignClient.dto.WeatherParams;
 import org.pknu.weather.feignClient.utils.ExtraWeatherApiUtils;
-import org.pknu.weather.feignClient.utils.WeatherApiUtils;
+import org.pknu.weather.feignClient.utils.WeatherFeignClientUtils;
 import org.pknu.weather.repository.ExtraWeatherRepository;
 import org.pknu.weather.repository.LocationRepository;
 import org.pknu.weather.repository.MemberRepository;
 import org.pknu.weather.repository.WeatherRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import static org.pknu.weather.dto.converter.ExtraWeatherConverter.toExtraWeather;
-import static org.pknu.weather.dto.converter.ExtraWeatherConverter.toExtraWeatherInfo;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
 public class WeatherService {
-    private final WeatherFeignClient weatherFeignClient;
+    private final WeatherFeignClientUtils weatherFeignClientUtils;
     private final WeatherRepository weatherRepository;
     private final ExtraWeatherRepository extraWeatherRepository;
     private final MemberRepository memberRepository;
     private final ExtraWeatherApiUtils extraWeatherApiUtils;
     private final LocationRepository locationRepository;
-    @Autowired
-    private EntityManager em;
-
-    @Value("${api.weather.service-key}")
-    private String weatherServiceKey;
-
-    /**
-     * 사용자의 위도 경도 및 기타 정보를 받아와 weather로 반환한다.
-     *
-     * @return now ~ 24 시간의 Wether 엔티티를 담고있는 List
-     * @Location 사용자 위치 엔티티
-     */
-    public List<Weather> getVillageShortTermForecast(Location location) {
-        float lon = location.getLongitude().floatValue();
-        float lat = location.getLatitude().floatValue();
-
-        PointDTO pointDTO = GeometryUtils.coordinateToPoint(lon, lat);
-        String date = DateTimeFormatter.getFormattedBaseDate();
-        String time = DateTimeFormatter.getFormattedBaseTime();
-
-        WeatherParams weatherParams = WeatherParamsFactory.create(weatherServiceKey, date, time, pointDTO);
-
-        WeatherApiResponse weatherApiResponse = weatherFeignClient.getVillageShortTermForecast(weatherParams);
-        List<WeatherApiResponse.Response.Body.Items.Item> itemList = weatherApiResponse.getResponse()
-                .getBody()
-                .getItems()
-                .getItemList();
-
-        return WeatherApiUtils.responseProcess(itemList, date, time);
-    }
+    private final WeatherWriteService weatherWriteService;
 
     /**
      * TODO: 성능 개선 필요
@@ -101,7 +62,7 @@ public class WeatherService {
      */
     @Transactional
     public List<Weather> saveWeathers(Location location) {
-        List<Weather> values = getVillageShortTermForecast(location);
+        List<Weather> values = weatherFeignClientUtils.getVillageShortTermForecast(location);
         List<Weather> weatherList = new ArrayList<>(values);
 
         weatherList.forEach(w -> w.addLocation(location));
@@ -111,39 +72,24 @@ public class WeatherService {
     /**
      * 날씨 정보를 저장합니다. 비동기적으로 동작합니다.
      *
-     * @param loc      member.getLocation()
+     * @param location member.getLocation()
      * @param forecast 공공데이터 API에서 받아온 단기날씨예보 값 list
      */
     @Async("threadPoolTaskExecutor")
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public void saveWeathersAsync(Location loc, List<Weather> forecast) {
-        Location location = locationRepository.safeFindById(loc.getId());
-        List<Weather> weatherList = new ArrayList<>(forecast);
-
-        weatherList.forEach(w -> w.addLocation(location));
-        weatherRepository.saveAll(weatherList);
+    public void saveWeathersAsync(Location location, List<Weather> forecast) {
+        weatherWriteService.saveWeathersAsync(location, forecast);
     }
 
     /**
      * 단기 날씨 예보 API가 3시간 마다 갱신되기 때문에, 날씨 데이터 갱신을 위한 메서드
      *
-     * @param loc API를 호출한 사용자의 Location 엔티티
+     * @param location API를 호출한 사용자의 Location 엔티티
      * @return 해당 위치의 날씨 데이터 List
      */
     @Async("threadPoolTaskExecutor")
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public void updateWeathersAsync(Location loc) {
-        Location location = locationRepository.safeFindById(loc.getId());
-        weatherRepository.deleteAllByLocation(location);
-
-        List<Weather> newWeathers = getVillageShortTermForecast(location).stream()
-                .toList();
-
-        newWeathers.forEach(weather -> {
-            weather.addLocation(location);
-        });
-
-        weatherRepository.saveAll(newWeathers);
+    public void updateWeathersAsync(Location location) {
+        weatherWriteService.updateWeathersAsync(location);
     }
 
     /**
